Move mottipets motifs
The motifs are similar to 'Mappings'. Multiple versions/settings can co-exists
next to each other no problem.
