--- CodeKlavier/motifs_interface.py
@@ -1,43 +1,0 @@
-"""
-
-(hradcoded) Motifs for Motippets
-
-"""
-#comment 
-class Motifs(object):
-    
-    def motif_1(self):
-        return [30, 35, 36, 37, 38, 39, 40, 41, 42, 43, 42, 41, 40, 39, 38, 37, 36, 35]
-    
-    def motif_2(self):
-        return [63,72,60,68,51,44,60,63,61,58,55,49]
-    
-    def mini_motif_1_low(self):
-        return [36,43,44,39,38]
-    
-    def mini_motif_2_low(self):
-        return [26,32,35,38]   
-    
-    def mini_motif_1_mid(self):
-        return [60,67,68,63,62]
-    
-    def mini_motif_2_mid(self):
-        return [50,56,59,62]    
-    
-    def mini_motif_1_hi(self):
-        return [84,91,92,87,86]
-    
-    def mini_motif_2_hi(self):
-        return [86,92,95,98]     
-    
-    def conditional_1(self):
-        return [45,43,46,45,43,41,43,45]
-    
-    def conditional_2(self):
-        return [36,31,29,26,28,35,38,33,24,26,31,28,33,21,23,29]
-    
-    def conditional_result_1(self):
-        return [60,62,66,70,69,67]
-    
-    def conditional_result_2(self):
-        return [87,80,78,68,75,79]
